\name{qRNASeq-package}
\alias{qRNASeq-package}
\alias{qRNASeq}
\docType{package}
\title{
  Quantify RNA-Seq data
}
\description{
  A package to quantify RNA-seq data in different level.
}
\details{
\tabular{ll}{
Package: \tab qRNASeq\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2012-03-14\cr
License: \tab GPL\cr
}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}
\author{
Qiang Hu, Li Yan, Dan Wang, Song Liu

Maintainer: Qiang Hu <huqmail@gmail.com>
}
\references{
%~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ RNA-Seq }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~
}
